requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pillow>=9.0.0
tqdm>=4.60.0
pyyaml>=6.0
gradio>=4.0.0
folium>=0.14.0
wandb>=0.15.0
